requirements.txt
sentence-transformers==2.7.0
transformers==4.41.2
faiss-cpu==1.7.4
numpy==1.26.4
flask==3.0.3
pypdf==4.2.0